Changing SantiRule semantics
SantiRules no longer allow you to specify the name of an operation as a key and the argument to that operation as the value of that key. You must now specify the operation name as the value of the key `op` and the argument as the value of the key `arg`. For example, previously, you could write:

```js
{
  select: 'span',
  allowStyles: ['color']
}
```

As a SantiRule. Now, you must write it as:

```js
{
  select: 'span',
  op: 'allowStyles',
  arg: ['color']
}
```

This change was made to reduce ambiguity when reading and processing SantiRules. This comes at the cost of brevity, which is a sacrifice I have decided is worthwhile.

SantiRules now have a `defaultSelect` property, which works like the `select` property, but won't overwrite an existing selection made earlier in execution of a ruleset. This is useful for giving rules default behavior even when used without context.

Finally, SantiRules also now have a `name` property, which is currently only used for making error messages more descriptive. Since SantiRules can't be arbitrarily stringified, this makes debugging easier. If you are not sure which exact rule is causing an error, give some rules `name`s and see which one appears.